% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEGs_Fct.R
\name{find_DEGs}
\alias{find_DEGs}
\title{find_DEGs}
\usage{
find_DEGs(x)
}
\arguments{
\item{x}{A data frame of false discovery rates (FDRs) or p values.}
}
\value{
A data frame with the same dimensions as the input dataframe, containing 0s and 1s.
}
\description{
Creates a data frame of 0s and 1s to represent differential expression based on a specified FDR threshold.
}
\details{
FDRs < 0.2 are assigned a 1 to indicate differential gene expression. FDRs > 0.2 are assigned a 0 to indicate no differential gene expression.
}
