% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_DEs_Fct.R
\name{label_DEs}
\alias{label_DEs}
\title{Label_DEs}
\usage{
label_DEs(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
interprets rowSums of find_DEGs function to create list of 0 (NDE) and 1s (DEs) and GeneIDs
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
