% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split_Dataframes_Fct.R
\name{split_df}
\alias{split_df}
\title{split_df}
\usage{
split_df(x, z = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{z}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
split data frame columns into separate dataframes, and adds suffix to dfs' names in case multiple Dfs that were created have the same names
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
