% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEGs_Fct.R
\name{find_DEGs}
\alias{find_DEGs}
\title{find_DEGs}
\usage{
find_DEGs(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
looks at dataframes of FDRs and creates matrix of same size as input df, labeling cells with 1s if FDR < 0.2 and 0 if FDR > 0.2
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
