% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_DEs_Fct.R
\name{label_DEs}
\alias{label_DEs}
\title{Label_DEs}
\usage{
label_DEs(x)
}
\arguments{
\item{x}{A data frame of 0s and 1s.}
}
\value{
An array with two columns, labels for 0 or 1, and GeneIDs that correspond to each label.
}
\description{
Uses the output from find_DEGs() to create an array with individual 0 or 1 labels per gene ID.
}
\details{
This function accepts a data frame containg 0s and 1s. For the purpose of this function, 1s indicate differentail gene expression and 0s indicate no differeential gene expression. This function takes the rowSums() of 0s and 1s per gene IDs, and assigns any gene with a sum > 1 with a 1 to indicate differential gene expression. Genee IDs with a sum of 0 are labeled with a 0 to represent no differential gene expression.
}
