% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split_Dataframes_Fct.R
\name{split_df}
\alias{split_df}
\title{split_df}
\usage{
split_df(x, z = NULL)
}
\arguments{
\item{x}{A data frame}

\item{z}{A label to be added as a suffix to the name of each created data frame.}
}
\value{
A data frame for each column from the input data frame.
}
\description{
Creates data frame out of each column in a given data frame.
}
\details{
This function creates a data frame in the Global Env. for each column in the input data frame. In case multiple columns have similar names, or you would like to keep track of the dataframe where these columns originated from, you can add a suffix to the end of the data frame name.
}
